refactor: move types from hook.ts and export them with aliases
