+++ src/components/employment/hooks.ts
@@ -1,22 +1,7 @@
 import { create } from "zustand";
+import { EmploymentAction, EmploymentState } from "./types";
 
-export type State = {
-  position: string;
-  employer: string;
-  startDate: string;
-  endDate: string;
-  description: string;
-};
-
-type Action = {
-  updatePosition: (position: State["position"]) => void;
-  updateEmployer: (employer: State["employer"]) => void;
-  updateStartDate: (sdate: State["startDate"]) => void;
-  updateEndDate: (edate: State["endDate"]) => void;
-  updateDescription: (desc: State["description"]) => void;
-};
-
-export const useStore = create<State & Action>()((set) => ({
+export const useStore = create<EmploymentState & EmploymentAction>()((set) => ({
   position: "",
   employer: "",
   startDate: "",
