+++ src/components/employment/types.ts
@@ -1,0 +1,17 @@
+type State = {
+  position: string;
+  employer: string;
+  startDate: string;
+  endDate: string;
+  description: string;
+};
+
+type Action = {
+  updatePosition: (position: State["position"]) => void;
+  updateEmployer: (employer: State["employer"]) => void;
+  updateStartDate: (sdate: State["startDate"]) => void;
+  updateEndDate: (edate: State["endDate"]) => void;
+  updateDescription: (desc: State["description"]) => void;
+};
+
+export { type State as EmploymentState, type Action as EmploymentAction };
